refactor(pdf-service): move createFooter, createPageNumber and createStamp out of domains folder
