+++ app/domains/fluggastrechte/services/pdf/sections/__test__/createFooter.test.ts
@@ -3,17 +3,25 @@ import {
   mockPdfKitDocument,
   mockPdfKitDocumentStructure,
 } from "tests/factories/mockPdfKit";
+import { createPageNumber } from "~/services/pdf/createPageNumber";
+import { createStamp } from "~/services/pdf/createStamp";
 import { createBankInformation } from "../createBankInformation";
 import { createFooter } from "../createFooter";
-import { createPageNumber } from "../createPageNumber";
-import { createStamp } from "../createStamp";
 
 vi.mock("../createBankInformation");
-vi.mock("../createPageNumber");
-vi.mock("../createStamp");
+vi.mock(
+  "/Users/rafaelfalk/Code/a2j-rechtsantragstelle/app/services/pdf/createPageNumber",
+  () => ({
+    createPageNumber: vi.fn(),
+  }),
+);
+vi.mock(
+  "/Users/rafaelfalk/Code/a2j-rechtsantragstelle/app/services/pdf/createStamp",
+  () => ({
+    createStamp: vi.fn(),
+  }),
+);
 
-vi.mocked(createStamp).mockImplementation(() => vi.fn());
-vi.mocked(createPageNumber).mockImplementation(() => vi.fn());
 vi.mocked(createBankInformation).mockImplementation(() => vi.fn());
 
 describe("createFooter", () => {
